# Performance improvement of KCallableImpl.callBy