> // TODO here we can avoid extra arguments copy by allocating an array of precise size if
> // varargs are not present and win some performance

This PR is intended to resolve the TODO comments here.
(I am a little unsure if this improvement is really what the comment intended, since I did not understand the meaning of "if vararg are not present"...)

This PR is based on the following
https://github.com/JetBrains/kotlin/pull/4840

Because some processes were difficult to port directly, we omitted the caching of the default argument setting process.

---

The local benchmark results are as follows

**Before:**

```
Benchmark                                 Mode  Cnt  Score   Error   Units
CallByBenchmark.callByLite               thrpt    5  6.996 ± 0.115  ops/us
CallByBenchmark.callByLiteWithLookup     thrpt    5  5.019 ± 0.055  ops/us
CallByBenchmark.callByReflect            thrpt    5  5.868 ± 0.127  ops/us
CallByBenchmark.callByReflectWithLookup  thrpt    5  4.327 ± 0.031  ops/us
```

**After:**

```
Benchmark                                 Mode  Cnt  Score   Error   Units
CallByBenchmark.callByLite               thrpt    5  7.843 ± 0.218  ops/us
CallByBenchmark.callByLiteWithLookup     thrpt    5  5.606 ± 0.041  ops/us
CallByBenchmark.callByReflect            thrpt    5  5.386 ± 0.107  ops/us
CallByBenchmark.callByReflectWithLookup  thrpt    5  4.520 ± 0.121  ops/us
```

It was faster, or at least not slower.